mhhh, bitshift for part 2 seems to be ...weird???

+++ Rust/day15/src/main.rs
@@ -1,30 +1,30 @@
-const VALUE_A: usize = 16807;
-const VALUE_B: usize = 48271;
+const VALUE_A: i64 = 16807;
+const VALUE_B: i64 = 48271;
 
-const INPUT_A: usize = 512;
-const INPUT_B: usize = 191;
+const INPUT_A: i64 = 512;
+const INPUT_B: i64 = 191;
 
-const DIVIDE_A: usize = 4;
-const DIVIDE_B: usize = 8;
+const DIVIDE_A: i64 = 4;
+const DIVIDE_B: i64 = 8;
 
-const DIVIDER: usize = 2147483647;
+const DIVIDER: i64 = 2147483647;
 
 struct Generator {
-    value: usize,
-    multiplier: usize,
+    value: i64,
+    multiplier: i64,
 }
 
 impl Generator {
-    fn new(value: usize, multiplier: usize) -> Generator {
+    fn new(value: i64, multiplier: i64) -> Generator {
         Generator {
             value: value,
             multiplier: multiplier,
         }
     }
 }
-//maybe swap the remain division.
+
 impl Iterator for Generator {
-    type Item = usize;
+    type Item = i64;
 
     fn next(&mut self) -> Option<Self::Item> {
         self.value *= self.multiplier;
@@ -40,6 +40,8 @@ fn part1() -> usize {
     generator_a.zip(generator_b).take(40_000_000).filter(|&(next_a, next_b)| (next_a ^ next_b) % 65536 == 0).count()
 }
 
+//maybe add 'item / (item >> 2) == DIVIDE_A', and 'item / (item >> 3) == DEVIDE_B'
+//however, this does not seem to work.
 fn part2() -> usize {
     let generator_a = Generator::new(INPUT_A, VALUE_A);
     let generator_b = Generator::new(INPUT_B, VALUE_B);
@@ -51,7 +53,6 @@ fn part2() -> usize {
         .filter(|&(next_a, next_b)| (next_a ^ next_b) % 65536 == 0).count()
 }
 fn main() {
-   //let (part1, part2) = rayon::join(||part1(), || part2());
    let part1 = part1();
    let part2 = part2();
    println!("part 1: {}", part1);
